feat: DealWallet 테이블 추가 및 스왑머니 사용 여부에 따른거래 수락 시 지갑에서 스왑머니 출금 입금 기능 구현, 거래 인수 시 각 지갑에 DealWallet에 있던 금액 입금

+++ src/main/java/piglin/swapswap/domain/deal/dto/request/DealCreateRequestDto.java
@@ -5,9 +5,9 @@ import jakarta.validation.constraints.PositiveOrZero;
 import java.util.List;
 public record DealCreateRequestDto (
         @PositiveOrZero
-        int firstExtraFee,
+        Long firstExtraFee,
         @PositiveOrZero
-        int secondExtraFee,
+        Long secondExtraFee,
         List<Long> firstPostIdList,
         List<Long> secondPostIdList,
         @NotNull
